Add Gender and Level Models and Update Student Model
Added new models for Gender and Level in the models.py file. These models represent the gender and level of a student in the database. Each model has an id and a name field. The Level model also has an abbreviation field.

Updated the Student model to include foreign key relationships to the Gender and Level models. This allows each student to have an associated gender and level.

Updated the list_students.html template to display the name of the student's gender and level instead of their id. This makes the student list more readable and user-friendly.