Add sorting by gender and level in student list
Updated the student list view to allow sorting by gender and level. This was achieved by modifying the query in the list_students function in students.py to join with the Gender and Level models when the sort parameter is 'gender.gender_name' or 'level.level_name'. Also updated the list_students.html template to include links for sorting by these fields.