Add student_id to DormRoom model and update student assignment logic
This commit adds a new column `student_id` to the `DormRoom` model in the database. This column will be used to store the ID of the student assigned to the room.

In the `students.py` file, the logic for assigning housing to a student has been updated to use the new `dorm_room2` relationship instead of `dorm_room_2`. This ensures that the correct relationship is used when assigning a room to a student.

A new migration file `c86b293f23a8_add_student_id_to_dormroom_model.py` has been created to add the `student_id` column to the `DormRoom` model.

Overall, this commit improves the functionality related to student housing assignment.

DORM ROOM STUFF WORKING RIGHT!!!