Add dormitory and room selection to student edit page
This commit introduces the ability for users to assign a dormitory and room to a student from the student edit page. A new dropdown has been added to select the dormitory, and upon selection, a second dropdown is dynamically populated with the rooms available in that dormitory. This is achieved through a new AJAX request to a new route '/get_rooms' which returns a list of rooms for a given dormitory. The selected room can then be assigned to the student. This feature enhances the user experience by allowing housing assignments to be made directly from the student edit page.