Add validation for room assignment in dorms
Added a check to ensure that a valid room is selected when assigning students to rooms in dorms. If an invalid room is selected, a flash message is displayed to the user. Also, added a check to ensure that the number of students assigned to a room does not exceed the room's capacity. If the capacity is exceeded, a flash message is displayed to the user. The flash messages are displayed in the assign_room.html template.